% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{chooseCoherentGeneNames}
\alias{chooseCoherentGeneNames}
\title{Choose coherent gene names}
\usage{
chooseCoherentGeneNames(in_df, their_data)
}
\arguments{
\item{in_df}{input data frame with column 'gene_name'}

\item{their_data}{reference data frame with column 'Gene_pSite'}
}
\description{
Choose coherent gene names
}
