% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getPhosphoSiteId}
\alias{getPhosphoSiteId}
\title{Create phosphosite id / Gene_pSite}
\usage{
getPhosphoSiteId(gene_name, string, seq, rm_pattern = "(ox)")
}
\arguments{
\item{gene_name}{vector of charaters indicating gene names}

\item{string}{vector of characters indicating the measured 
peptide sequence with indicated phosphosite "p"}

\item{seq}{vector of characters with protein sequence 
corresponding to gene names}

\item{rm_pattern}{character string to be removed from
modified sequence string}
}
\description{
Create phosphosite id / Gene_pSite
}
