% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getProb}
\alias{getProb}
\title{Get localization site probability from MaxQuant file}
\usage{
getProb(x, p_thres = 0.75)
}
\arguments{
\item{x}{column name of data frame or vector including
sequence including positional probability as created
by MaxQuant}

\item{p_thres}{probability threshold for good phosphosite
localization, deafult is 0.75}
}
\description{
Get localization site probability from MaxQuant file
}
