% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{lowerFn}
\alias{lowerFn}
\title{Lower Functions for GGally}
\usage{
lowerFn(data, mapping, dot.size)
}
\arguments{
\item{data}{data frame imported from ggpairs}

\item{mapping}{variable imported from ggpairs}

\item{dot.size}{size of geom_point}
}
\description{
Lower Functions for GGally
}
