% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plotDiffPhosphoSpline}
\alias{plotDiffPhosphoSpline}
\title{Fit differentil melting curves fitted by Splines}
\usage{
plotDiffPhosphoSpline(g_name, p_seq, spline_df = 4,
  predict_vec = seq(35, 70, by = 0.1), nbf_df, phospho_df, Gene_pSite)
}
\description{
Fit differentil melting curves fitted by Splines
}
