% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plotDiffPhosphoMeltcurve}
\alias{plotDiffPhosphoMeltcurve}
\title{Plot differential melting curves}
\usage{
plotDiffPhosphoMeltcurve(
  g_name,
  p_seq,
  predict_vec = seq(35, 70, by = 0.1),
  nbf_df,
  phospho_df,
  Gene_pSite,
  in_ylim = c(0, 1.25)
)
}
\arguments{
\item{g_name}{gene name}

\item{p_seq}{phophopeptide sequence}

\item{predict_vec}{vector of temperatures to predict 
melting curve fits for}

\item{nbf_df}{data frame of unmodified qunantitative data}

\item{phospho_df}{data frame of phosphopeptide quantitative data}

\item{Gene_pSite}{phosphosite id to be displayed as title}

\item{in_ylim}{limits for y-axis}
}
\description{
Plot differential melting curves
}
